+++ src/main/java/data/FileParser.java
@@ -18,32 +18,8 @@ public class FileParser {
     this.basePath = filepath;
   }
 
-  /** Map each vendor ID to an array of account IDs.
-   * @return JSON dictionary
-   */
-  public String getConfigMap() {
-    ArrayList<String> vendorIDs = getVendorIDs();
-    HashMap<String, ArrayList<String>> map = new HashMap<>();
 
-    for (String vendor : vendorIDs) {
-      ArrayList<String> accounts = getAccountIDs(vendor);
-      map.put(vendor, accounts);
-    }
 
-    Gson gson = new Gson();
-    String json = gson.toJson(map);
-    return json;
-  }
 
-  /** Return list of vendor IDs that exist in the filesystem.*/
-  private ArrayList<String> getVendorIDs() {
-    File root = new File(basePath);
-    return new ArrayList<>(Arrays.asList(Objects.requireNonNull(root.list())));
-  }
-
-  /** TODO Return list of account IDs corresponding to a given vendor. */
-  private ArrayList<String> getAccountIDs(String vendorID) {
-    return new ArrayList<>(Arrays.asList("ECHO", "FOXTROT", "GOLF", "HOTEL"));
-  }
 }
